Keep the stolen/impounded map circle over a fixed ground area
circle-radius is in screen pixels, so the old ad-hoc curve
(stops [[5, 5], [16, 240]]) grew it only 48x across 11 zoom levels while
the map scale changed 2048x — the circle covered progressively more
ground as you zoomed out. Derive the pixel radius from a real ground
radius instead, interpolating base-2 so it doubles every zoom level.

The opaque radiusBase curve constant becomes radiusMeters. Values are
picked to match the old on-screen size at the default zoom 13, so the
initial view is unchanged.

Also extracts the MapLibre + PMTiles CDN loader into utils/maplibre,
matching sethherr/parking-notification-map-pin.
