+++ app/components/registrations/show/map/component.html.erb
@@ -3,7 +3,7 @@
   data-registrations--show--map-style-url-value="<%= MAPS_STYLE_URL %>"
   data-registrations--show--map-latitude-value="<%= @latitude %>"
   data-registrations--show--map-longitude-value="<%= @longitude %>"
-  data-registrations--show--map-radius-base-value="<%= radius_base %>"
+  data-registrations--show--map-radius-meters-value="<%= radius_meters %>"
   data-registrations--show--map-point-value="<%= @point %>"
   class="
     tw:relative tw:mt-3 tw:mb-4 tw:h-80 tw:w-full tw:overflow-hidden
